Win32-network: reorganise modules
