+++ Win32-network/test/Main.hs
@@ -11,11 +11,11 @@ module Main (main) where
 #if defined(mingw32_HOST_OS)
 import           Test.Tasty
 
-import qualified Test.Async
+import qualified Test.Async.Handle
 
 main :: IO ()
 main = defaultMain $ testGroup "Win32"
-  [ Test.Async.tests
+  [ Test.Async.Handle.tests
   ]
 #else
 main :: IO ()
